Deck building approach
Sorry for the messy code.

This commit make it so that the MCTS AI tries to build a deck first.
Player One get's to choose between a fixed set of cards that I (Ragowit)
have in my collection.
Player Two can choose from the entire game library.
Player Two starts by picking a class, and then continues on with being
the first one to pick a card.

Other changes include that the AI tries every position for a minion
summon and tries all different targets for battlecries as well.