Changed the card picking phase
Previously so did the AI first pick a hero, and then it started to
alternate card picking between player two ("the opponent") and player
one ("me"). This took 61 turns before a deck for "me" was complete and
was specific against the hero that the AI had chosen in the very first
turn.

Now, the AI first choose 30 cards for player one ("me"), then chooses a
hero for player two and its 30 cards. I believe that this approach is
better since its 1) not targeting a specific hero, and 2) only takes 30
turns to create a deck for "me" - allowing me to increase the thinking
time of AI while not increase the time spent for a created deck.

The thinking time has been increased to 1000 seconds per turn, from 480
seconds. A complete deck will be done in 500 minutes (8 hours and 20
minutes).