More adjacent cards to look out for
Added Piloted Shredder and Piloted Sky Golem, since the first one can
get Sunfury Protector and the second one can get Defender of Argus. They
don't affect the board when summoned by these cards, but if other cards
exists that returns them to hand they do affect the placement of
minions.

It would probably be better if you checked for Piloted Shredder/Piloted
Sky Golem AND a bounce card, but that's something for the future.